Added the ability to append and prepend text on links. IE icons
