+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "laravel-menu"]
+	path = laravel-menu
+	url = git@github.com:ccsliinc/laravel-menu.git
