+++ src/Traits/Conditions.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace Spatie\Menu\Traits;
+
+
+trait Conditions {
+    /**
+     * @param $conditional
+     * @return bool
+     */
+    protected function resolveCondition($conditional)
+    {
+        return is_callable($conditional) ? $conditional() : $conditional;
+    }
+}
